requirements.txt
asgiref==3.8.1
certifi==2024.7.4
Django==5.0.6
djangogrpcframework==0.2.1
djangorestframework==3.15.2
drf-yasg==1.21.7
grpcio==1.64.1
grpcio-tools==1.64.1
inflection==0.5.1
multidict==6.0.5
packaging==24.1
pillow==10.3.0
protobuf==5.27.1
psycopg2-binary==2.9.9
pulsar-client==3.5.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
setuptools==72.1.0
sqlparse==0.5.0
uritemplate==4.1.1